Initial commit to modify bundle content
This adds net-new steps required to modify a bundle image for Konflux.
This type of flow can be added into any existing processing/tooling for
bundle generation.

The source CSV already contains the relatedImages section so this
tooling does not need to take that into account. If needed, a tool like
https://github.com/containerbuildsystem/operator-manifest can be used to
generate the relatedImage specification.

By including the pullspecs of all images that will be referenced in the
CSV in a single file, it will enable these to be easily nudged in the
future after new builds have completed.

Signed-off-by: arewm <arewm@users.noreply.github.com>