The index image generation was being skipped
Since the `build-image-index` depends on the results from each of the
architecture-specific builds, the task (and all resulting dependant
tasks) will be skipped if any of these are absent.

Using the intended workflow of enabling users to specify the
architectures to run with PipelineRun params will require this model to
change. The PR https://github.com/konflux-ci/build-definitions/pull/1111
is a potentially promising way to resolve this issue.

Signed-off-by: arewm <arewm@users.noreply.github.com>